modificacion de consultas
